% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2.sim.849.R
\name{hw2.sim.849}
\alias{hw2.sim.849}
\title{A function to do question 2 from homework 2 in STAT 849}
\usage{
hw2.sim.849(h0.mean, num.samp, num.reps, samp.dist, times.scalar = FALSE,
  alpha = 0.05, ...)
}
\arguments{
\item{h0.mean}{The mean under H0}

\item{num.samp}{The number of samples in each repetition}

\item{num.reps}{The number of repetitions}

\item{samp.dist}{The distribution to sample from, either "normal", "chisq", or "t"}

\item{times.scalar}{This is added for part g. If true, the function randomly samples a 5 or 10 for each repetition and multiples that by the first observation in that repetition's sample.}

\item{alpha}{Alpha level for test}

\item{...}{Additional options to specify for the sampling distribution}
}
\description{
This function simulates data from  a distribution and performs a one sample T-test then outputs coverage probability and type 1 error rate
}
\keyword{Classwork,}
\keyword{STAT849,}
\keyword{simulations}
