% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imaging.ROC.R
\name{imaging.ROC}
\alias{imaging.ROC}
\title{A function to create ROC plots from a list of prediction images for segmentation}
\usage{
imaging.ROC(preds.list, mask.list, y.list, vec.id, type, group.list, max.fpr)
}
\arguments{
\item{preds.list}{A list of prediction images as NIfTI objects for different subjects (can be of length 1)}

\item{mask.list}{A list of brain masks as NIfTI objects in the same order as preds.list}

\item{y.list}{A list of the gold standard segmentations as NIfTI objects in the same order as preds.list.}

\item{vec.id}{A vector of ID's (as characters) corresponding to preds.list}

\item{type}{There are 3 options: "subject", "total", or "group". "subject" creates a seperate ROC for each subject, "total" creates one overall ROC, and by group makes a seperate ROC for each group as determined by the group.list argument.}

\item{group.list}{A list of sub-lists. This is only necessary if type="group". This should be a list of sublists where each sublist is the IDs of the subjects wanted in each group. For example, list(list("1","2"),list("3","4")) would create two ROCs, one with subjects "1" and "2" and the other with subjects "3" and "4".}

\item{max.fpr}{A scalar between (0,1). Use this to create partial ROCs instead of full ROCs. The ROC will be cut-off at the max false positve rate (max.fpr) specified.}
}
\description{
This function creates different types of ROC plots from prediction images for segmentation. It also calculates AUC and returns the data.frame used for the plot.
}
\examples{
imaging.ROC()
}
\keyword{AUC,}
\keyword{NIfTI}
\keyword{ROC,}
