% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img.stand.R
\name{img.stand}
\alias{img.stand}
\title{A function to standardize NIfTI objects}
\usage{
img.stand(image, mask, trim)
}
\arguments{
\item{image}{NIfTI object, image to standardize.}

\item{mask}{NIfTI object, brain mask of image.}

\item{trim}{scalar between (0,1), percent of trimming to do from the highest and lowest intensities}
}
\description{
This function allows you to standardize NIfTI objects. It also allows for you to do a trimmed standardizating (subtract trimmed mean, divide by trimmed standard deviation)
}
\examples{
img.stand()
}
\keyword{MRI,}
\keyword{NIfTI,}
\keyword{normalization,}
\keyword{trimmed}
