% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot objects}

\item{plotlist}{a list of ggplot objects}

\item{cols}{Number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.
#' @keywords ggplot, ggplot2}
}
\description{
This function allows you to plot multiple ggplots on the same plot. The ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)
}
\examples{
multiplot()
}
